Insert now completed fully functional.
B+ tree insert now works for all cases, there seems to be a little bug with the keys array not setting NIL some unused spots but seems harmless as everything is driven by node size which seems to be properly working.
Renamed the search and insert functions.
Removed test code, I will rewrite it with the new constructs.