Changed internal representation into an struct.
Changed the nodes to a struct to take advantage of the getter methods, provided setters to acompany them.
Fixed a bug in the key promotion.
Made the node splitter function to set the pointer to the next leaf in the chain, as to be able to traverse the leaves horizontally for a ranged search.
Changed the naming of several functions.