bplustree-delete works.
Deleting now works for all cases, when possible transfers records from neighboring nodes to balance the tree, if not possible it merges two nodes to rebalance, if this happen the internal node loses a key, thus the same process happens again, going all the way to the top, with the posibility of the root losing a level.
Removed the depth field, hard to mantain for every node.