Backward-compatible extensions.
bplustree-insert now takes an optional KEY parameter. If included, it is used instead of calling the key function on the record.

New functions: bplustree-empty-p, bplustree-search-next, bplustree-search-prev, bplustree-traverse-node-with-keys

A tree now caches the last (key record node) returned from bplustree-search, bplustree-search-next, or bplustree-search-prev, and uses it to optimize calls of those functions with a repeated key. This is a big win for the expected uses of bplustree-search-next and bplustree-search-prev.

To support bplustree-search-prev, the leaf nodes are now doubly-linked.